Add functionality to receive js from server
When sending an EMA model to the server it can respond with js code that
is then added to the DOM tree and executed. The rational behind this is
to have the server compile the EMA model to WebAssembly which is
displayed in the original website then.

As a test, there is a random image painted where the model will be
visible later.