Add multiple deletion attempts
The filesystem may be just slow the first time we attempt to delete a
directory. Even though delete was called on every file recursively in a
directory, it sometimes fails to delete saying the directory is not
empty.

Therefore, we wait a reasonable amount of time and try again before
giving up and throwing the exception.